redistribute variables and move buying/selling func to modal confirm
